server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.14
fastapi-cli==0.0.7
flatbuffers==25.2.10
frozenlist==1.7.0
future==1.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
google-genai==1.24.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
numba==0.61.2
numpy==1.26.4
onnxruntime==1.20.1
orjson==3.10.18
packaging==25.0
pillow==11.1.0
pipecat-ai==0.0.91
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.19.2
pyloudnorm==0.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
resampy==0.4.3
rich==14.0.0
rich-toolkit==0.14.8
rsa==4.9.1
scipy==1.16.0
shellingham==1.5.4
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==13.1
yarl==1.20.1
